Create rodcuttingproblem1.java
Problem Statement
Given a rod of length ‘n’ units. The rod can be cut into different sizes and each size has a cost associated with it. Determine the maximum cost obtained by cutting the rod and selling its pieces.
Note:
The sizes will range from 1 to ‘n’ and will be integers.
The sum of pieces cut should be equal to ‘n’.
Consider 1-based indexing.
Input Format:
The first line of input contains an integer ‘T’ denoting the number of test cases.
The next ‘2*T’ lines represent the ‘T’ test cases.

The first line of each test case contains an integer ‘n’ denoting the length of the rod.
The second line of each test case contains a vector ’A’, of size ‘n’ representing the cost of different lengths, where each index of the array is the sub-length and the element at that index is the cost for that sub-length.

Note: Since 1-based indexing is considered, the 0th index of the vector will represent sub-length 1 of the rod. Hence the (n-1)th index would represent cost for the length ‘n’. 
Output Format
For each test case, return the maximum cost obtained by selling the pieces.
Constraints:
1 <= T <= 50
1 <= n <= 100
1 <= A[i] <= 100

Where ‘T’ is the total number of test cases, ‘n’ denotes the length of the rod, and ‘A[i]’ is the cost of sub-length.

Time limit: 1 second
Sample Input 1:
2
5
2 5 7 8 10
8
3 5 8 9 10 17 17 20
Sample Output 1:
12
24