Create covidspread.java
Problem Statement
You are given a city which contains N x M houses, where each house can have one of the following three conditions :
1. The value ‘0’ represents an empty house,
2. The value ‘1’ represents a non-infected person,
3. The value ‘2’ represents an infected person.
It takes one day to propagate the infection from an infected house to its adjacent (Front, Back, Left, Right) non-empty and non-infected house. An empty house can only break the line of propagation of infection.
You need to return the minimum number of days Covid will take to infect each and every house in the city. And for the God’s sake if this is impossible, return -1 instead.
Input Format :
The first line contains a single integer ‘T’ denoting the number of test cases.

The first line of each test case will contain two integers ‘N’ and ‘M’ denoting the number of rows and columns, respectively.

Next ‘N’ lines contain ‘M’ space-separated integers each denoting the conditions of the house in the city as above explained.
Output Format :
For each test case, print the minimum number of days Covid will take to infect each and every person in the city. If this is impossible, print -1 instead.

The output of every test case will be printed in a separate line.
Note :
You do not need to print anything; It has already been taken care of.
Constraints :
1 <= T <= 50
1 <= N, M <= 100
0 <= houses[i][j] <= 2

Where ‘T’ is the number of test cases.
Where 'N' is the number of rows in the given matrix, and 'M' is the number of columns in the given city.
And, houses[i][j] denotes the condition of (i,j) house in the city.

Time limit: 1 sec
Sample Input 1 :
2
2 2
2 2
0 1
2 3
1 0 1
2 1 0
Sample Output 1 :
1
-1